plotting with gmt works

+++ script/gmtplot.sh
@@ -8,9 +8,11 @@ USE_YMD_FORMAT=0
 REMOVE_OUTLIERS=0
 ##  current pid
 pid=$(echo $$)
+##  number of ticks on Y-axis
+Y_TICKS_NR=6
 ## output file
 ps=koko.ps
-rm $ps > /dev/null
+rm $ps 2>/dev/null
 
 while test $# -gt 0
 do
@@ -56,7 +58,11 @@ then
     gmt set FORMAT_DATE_MAP o
     gmt set FORMAT_TIME_PRIMARY_MAP abbreviated
     gmt set FORMAT_DATE_OUT yyyy-o-dd
-    ./mjd2ymdhms.py -f ${cts_file} -s 1 -d ',' > ${cts_file}.tmp
+    if ! ./mjd2ymdhms.py -f ${cts_file} -s 1 -d ',' > ${cts_file}.tmp
+    then
+        echoerr "[ERROR] Failed to transform dates in input file."
+        exit 1
+    fi
     mv ${cts_file}.tmp ${cts_file}
     wesn=( $(gmt info -fT -I0.01 -C @"${cts_file}" --FORMAT_DATE_IN=yyyy-mm-dd -h --IO_N_HEADER_RECS=1) )
 else
@@ -73,33 +79,53 @@ Ry="-R${wesn[0]}/${wesn[1]}/${ymin}/${ymax}"
 zmin="${wesn[10]}"
 zmax="${wesn[11]}"
 Rz="-R${wesn[0]}/${wesn[1]}/${zmin}/${zmax}"
+
+# x/y limits array
 R=("${Rx}" "${Ry}" "${Rz}")
+# file column array
+Y=("2" "5" "8")
+# y annotation every
+tcks1=$(./annotevery.py ${xmin} ${xmax} ${Y_TICKS_NR})
+tcks2=$(./annotevery.py ${ymin} ${ymax} ${Y_TICKS_NR})
+tcks3=$(./annotevery.py ${zmin} ${zmax} ${Y_TICKS_NR})
+YAI=("${tcks1}" "${tcks2}" "${tcks3}")
+# x annotation ticks
+XAT=("WseN" "Wsen" "WSen")
 
-gmt set FONT_ANNOT_PRIMARY +10p
+gmt set FONT_ANNOT_PRIMARY +8p
 gmt set PS_CHAR_ENCODING ISOLatin1+
-echo $Rx
 
-it=0
-for comp in north east up
+it=2
+for comp in up east north
 do
-    Yshift=$(( ${it} * 3 ))
+    yshift="2.2"
+    ycol="${Y[$it]}"
+    if test "${it}" -eq 2
+    then
+        Osw=
+        Ysh=
+    else
+        Osw="-O"
+        Ysh="-Y${yshift}i"
+    fi
     if test "${USE_YMD_FORMAT}" -eq 1
     then
-        gmt psbasemap "${R[$it]}" -JX9i/2i -O -K -Bsx1Y -Bpxa6Of1o -Bpy0.05 \
-            -BWSen+t"Time Series"+glightgreen -Y${Yshift}i >> $ps
+        gmt psbasemap "${R[$it]}" -JX10i/2i -K ${Osw} \
+            -Bsx1Y -Bpxa6Of1o \
+            -Bpy"${YAI[$it]}"+l${comp} -B"${XAT[$it]}"+t"Time Series"+glightgreen \
+            ${Ysh} >> $ps
     else
-        gmt psbasemap "${R[$it]}" -JX9i/2i -O -K -Bsx100 -Bpy0.05 \
-            -BWSen+t"Time Series"+glightgreen -Y${Yshift}i >> $ps
+        gmt psbasemap "${R[$it]}" -JX10i/2i -K ${Osw} \
+            -Bsx100 \
+            -Bpy"${YAI[$it]}"+l${comp} -B"${XAT[$it]}"+t"Time Series"+glightgreen \
+            ${Ysh} >> $ps
     fi
 
     cat ${cts_file} | \
-        awk -F "," '{print $1,$2}' | \
+        awk -F "," -v y=$ycol '{print $1,$y}' | \
         gmt psxy "${R[$it]}" -J \
         -Wthin,red  -Gblack -Sc.05 \
         -h --IO_N_HEADER_RECS=1 -O -K >> $ps
 
-    let it=it+1
+    let it=it-1
 done
-cat ${cts_file} | head -2 | awk -F "," '{print $1,$2}' | gmt psxy "${R[2]}" -J -Wthin,red  -Gblack -Sc.05  -h --IO_N_HEADER_RECS=1 -O >> $ps
-
-echo "cts file was: ${cts_file}"
